@@ -1,13 +1,7 @@
 # Generated by make_width_table.py
 # wcwidth 0.2.6
 # Unicode 15.0.0
-import sys
-from typing import List, Tuple
-
-if sys.version_info < (3, 8):
-    from typing_extensions import Final
-else:
-    from typing import Final
+from typing import Final, List, Tuple
 
 WIDTH_TABLE: Final[Tuple[Tuple[int, int, int], ...]] = (
     (0, 0, 0),
